+++ src/test/java/org/springframework/samples/petclinic/owner/VisitControllerTests.java
@@ -29,7 +29,7 @@ import org.springframework.test.web.servlet.MockMvc;
 @WebMvcTest(VisitController.class)
 public class VisitControllerTests {
 
-    private static final int TEST_PET_ID = 1;
+    private static final long TEST_PET_ID = 1;
 
     @Autowired
     private MockMvc mockMvc;
